end: added flag for signing images
This adds support for having acbuild use gpg for you to sign the
produced aci. By default gpg is called with the following flags:

`--armor --yes --output mycoolapp.aci.asc --detach-sig mycoolapp.aci`

If any flags are provided to `end` after the path to the ACI, the
`--armor` and `--yes` flags are removed, and these provided flags are
prefixed to the list of flags passed to gpg.

There is also a `--user` flag on `end` that, when provided, will specify
the user (and potentially group) to run the `gpg` command as.