*: refactored library functions, added struct with common fields
The library was refactored so that all the exposed functions were
attached to a struct containing the information about the current
context. This makes using the library much cleaner.

The Registry struct had some of its fields renamed to be consistent with
the new struct in the library.

The functions for acquiring and releasing the lock were also moved from
acbuild/acbuild.go to the new struct.