run: --debug flag caused a panic with the Uncompress progress bar
The recent refactor resulted in the registry being correctly passed the
debug value, which exposed an issue with an additional progress bar that
was being drawn when uncompressing a dependency.

The panic was a combination of the logic drawing the bar passing the
ioprogress library a number that was too small for the total size, and
the ioprogress library not having a patch to be able to handle incorrect
numbers.

This commit removes the uncompressing bar, and updates the ioprogress
library.

Fixes https://github.com/appc/acbuild/issues/56