*: --modify was broken, didn't call command with different work context
Fixes https://github.com/appc/acbuild/issues/67

The --modify flag creates a new temp directory, and then uses that as a
the build context for a call to begin, the given command, write, and
end. It was failing to use the alternate work context for the given
command.

The temp directory's path used by the --modify flag is also now based on
a sha512 hash of the absolute path of the given ACI to modify. This aims
to prevent concurrent invocations of --modify on the same ACI.