*: fix uses of empty types.App resulting in invalid manifest
When presented with a manifest that lacked an App, many places in the
lib package would fill it in with a &types.App{}. This is incorrect, as
the user and group fields must be set in the struct for the manifest to
be valid. Without this fix, acbuild will error when asked to modify an
existing ACI in such a way that requires creating an App in the
manifest.