Added BeforeAfterFSDiffer. Will report changes in a directory.
For acbuild I want to walk the filetree to take note of the filesystem,
allow the user provided command to do its things, and then walk it again
to take note of deleted files. The SimpleFSDiffer will report changes
between two different directories at a given point in time. This
introduces the BeforeAfterFSDiffer, which will report changes in a
single directory between two different points in time.