begin: require first char to be / or . when starting from local file
So as to provide better error messages when a user attempts to start a
build from a local file and this file doesn't exist, the user must
provide a path to the file starting with '/' or '.'.

Additionally, begin now properly cleans up whatever context it started
to make if an error is returned.