copy: allow for copying from multiple sources at once
This commit allows for multiple sources for a copy to be specified, so
`acbuild copy` behaves a little more like `cp`. If there are multiple
sources, the destination must be a directory. If there are multiple
sources and the destination directory doesn't exist, the directory is
created.

This allows the * expansion from the shell, which makes copying all the
contents of a directory to a location in the ACI much easier.

For example: acbuild copy html/* /usr/share/nginx/html/