modify: don't require . or / for starting char, don't delete on failure
When the modify flag was used, it was requiring that the initial
character for the path provided start with a `.` or a `/`. Since the
modify flag can only be used on local images, this was unnecessary.

Also when using the modify flag and the write failed (e.g. due to
permissions issues), the original image was deleted. This commit changes
it so that the write happens to a different file from the original, and
on success replaces the original with the new image.