run: don't pass certain flags to systemd-nspawn for systemd < 209
The -q/--quiet flag and the --register flag were added to systemd-nspawn
in version 209, and some distros (like CentOS 7) ship with an older
version than that. This commit runs and parses `systemctl --version` to
determine the version number of systemd, and will only pass these flags
if the version is high enough.