run: set the PATH variable
Depending on the image, the PATH variable inside of a acbuild run
command could be propagated from something on the host. On normal
distros, this was mostly a non-issue. On NixOS, it would get set to
"/no-such-path:/sbin:/bin", which is definitely an issue.

This commit makes acbuild adhere to the AppC spec by setting the PATH
environment variable to the following:
"/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin".

Note that it only sets it on the systemd-nspawn command, it doesn't
add or modify the PATH variable in the manifest.