run: check that the binary exists, error if not, warn if symlink
acbuild run will now check if the given binary exists, and will print a
nice error message if it doesn't, as opposed to relying on
systemd-nspawn to inform the user.

Additionally on systems where systemd's version is <228, systemd may
error out when attempting to exec a symlink. This adds a check for this
case, and prints a warning.