copy: split --to-dir flag into separate subcommand
Copy's semantics are weird and I don't like them, but there are
technical reasons preventing `acbuild copy` from being reasonable. This
commit breaks the --to-dir flag on `acbuild copy` out into a separate
command, so that when users notice that `acbuild copy` doesn't do what
they expect it'll be a little more obvious that there are two modes of
operation here. Hiding this away in a flag, and therefore inside a help
page that wasn't at the top level, wasn't very user friendly.