*: exit with error code 3 when arguments are invalid
If the user calls acbuild with invalid subcommands or flags, and the
code for no subcommand is called, acbuild would up until now exit with
exit code 0. This can break scripts relying on the exit code to catch
errors.

This commit changes this such that acbuild will exit with code 3 when
called with invalid subcommands or flags.