begin: when provided file doesn't look like an ACI, it's a rootfs tarball
This commit modifies begin such that after a local file has been given
to acbuild and it's been extracted, acbuild will check for the existence
of a manifest and a rootfs. If either is missing, it's assumed that the
tarball acbuild just extracted is the rootfs for the container, and it
will move the files into the rootfs directory and create an empty
manifest.