run: add support for swappable engines
This commit revamps `lib/run.go` to rely on an interface for the actual
execution of the desired command, such that alternate ways to run a
command in the container can be easily added. A flag has been added to
the run command to select a non-default engine. The default (and
currently only) engine is systemd-nspawn.