godep/build: rkt to v1.2.0, unrewrite imports, tweak build & test scripts
Wanted to bump appc to 0.7.4, which is the result of bumping rkt to
1.2.0.

Also unrewrote imports, and tweaked the build and test scripts
appropriately.

The build and test scripts were also modified to work when called from
any directory, instead of only working when the current directory was
the acbuild repo.