begin: when on arm, look at self's endianness to pick a label
The runtime.GOARCH value isn't accepted by the AppC spec as a value for
the arch label in a manifest when on an arm platform. This commit tweaks
the begin command, when on arm, to get acbuild to correctly pick an
armv7 label based on the endianness of the machine it is running on.