engine/systemdnspawn: create /etc/machine-id for >= 230
Starting in systemd 230, systemd-nspawn requires the file
/etc/machine-id to exist in the container for some reason. There does
not appear to be a flag to disable this, and the commit message I found
for the systemd commit that I think causes this is rather vague as to
why this is necessary.

As a result, this commit creates /etc/machine-id before calling
systemd-nspawn and removes it afterwards.