*: added a script mode
This commit adds a script mode to acbuild, whereby build scripts can be
written in an acbuild DSL and then run. When acbuild is run in this way,
acbuild will exec itself with the debug flag, a work-path flag, and the
contents of each line. This lives under the "script" subcommand, but
this subcommand will be implied if the run acbuild has the name
"acbuild-script" (this was necessary to support /usr/bin/env shebangs).
The apache example has been rewritten to be in the acbuild DSL.

With a DSL mode various enhancements are now possible (like per-command
build caches), but the immediate benefit is the fact that acbuild can
enforce cleanup of builds that use the script mode.