Workaround: install missing Virtualbox Guest Additions in Xenial
As originally noted in
https://github.com/appc/acbuild/issues/234#issuecomment-238365738, it appears
the script `scripts/vagrant-install-go.sh` intends to be executed via the
synced folder functionality of Virtualbox, which requires the Virtualbox Guest
Additions to be installed in the guest VM. As noted, this is missing from the
Xenial image (see bug https://bugs.launchpad.net/cloud-images/+bug/1565985).
This implements a workaround for this bug, and is a first step towards fixing
the Vagrant-based option for building `acbuild` (see
https://github.com/appc/acbuild/issues/234#issuecomment-238393500 for why this
is just the first step).