Added support for OCI builds
This commit adds support to acbuild for building OCI compatible
container images. A new flag has been added to the `begin` command,
`--build-mode`, to specify when a build is started if it is for creating
an appc-compliant image or an OCI-compliant image.

The isolator, dependency, set-supp-groups, set-event-handler, and
set-name subcommands only work under the appc build mode.

All other subcommands work under both build modes, but obviously some
will have slightly different semantics.

A new subcommand named `layer` has been added. This subcommand only
works under oci build mode. It adds a new layer to the image, and future
file changes (via copy or run) will occur in the new layer.

A new flag, `--config`, has been added to the `cat-manifest` subcommand,
so that when the build mode is oci the config can be printed as well.